refactor: extract optimism receipts hashing logic to reth-optimism-consensus
