+++ crates/evm/execution-types/Cargo.toml
@@ -12,7 +12,6 @@ workspace = true
 
 [dependencies]
 reth-primitives.workspace = true
-reth-chainspec = { workspace = true, optional = true }
 reth-execution-errors.workspace = true
 reth-trie.workspace = true
 
@@ -27,6 +26,6 @@ alloy-eips.workspace = true
 
 [features]
 default = ["std"]
-optimism = ["dep:reth-chainspec"]
+optimism = []
 serde = ["dep:serde", "reth-trie/serde", "revm/serde"]
 std = []
