+++ crates/optimism/payload/Cargo.toml
@@ -29,6 +29,7 @@ reth-chain-state.workspace = true
 
 # op-reth
 reth-evm-optimism.workspace = true
+reth-optimism-consensus.workspace = true
 reth-optimism-forks.workspace = true
 
 # ethereum
